tests/wpt/css-tests/cssom-view-1_dev/xhtml1/cssom-getClientRects.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
    <meta charset="utf-8" />
    <title>CSS Test (CSSOM View): getClientRects of element outside DOM</title>
    <link href="http://chrisrebert.com" rel="author" title="Chris Rebert" />
    <link href="http://www.w3.org/TR/cssom-view/#dom-element-getclientrects" rel="help" />
    <meta content="dom" name="flags" />
    <meta content="Calling getClientRects on an element that is outside of the DOM (and therefore does not have an associated layout box) should result in an empty DOMRectList and should definitely not throw an error." name="assert" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        test(
            function () {
                var rectList = document.createElement('div').getClientRects();
                assert_equals(rectList.length, 0, "DOMRectList should be empty");
            },
            "getClientRects on a newly-created Element not yet inserted into the DOM should return an empty DOMRectList"
        );
    </script>
</head>
<body>

</body></html>
